+++ src/main/java/kitchenpos/common/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.common;
 
 public enum OrderStatus {
     COOKING, MEAL, COMPLETION
