+++ src/main/java/kitchenpos/common/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.common;
 
 import java.math.BigDecimal;
 import java.util.Objects;
