+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -7,7 +7,7 @@ import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import kitchenpos.domain.Price;
+import kitchenpos.common.Price;
 
 @Entity
 public class Menu {
