+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -3,7 +3,7 @@ package kitchenpos.order.application;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.table.repository.OrderTableRepository;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.common.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.menu.repository.MenuRepository;
 import kitchenpos.order.domain.Order;
