+++ src/main/java/kitchenpos/product/doamin/Product.java
@@ -6,7 +6,7 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import kitchenpos.domain.Price;
+import kitchenpos.common.Price;
 
 @Entity
 public class Product {
