+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.table.repository.OrderTableRepository;
 import kitchenpos.tablegroup.repository.TableGroupRepository;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.common.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTables;
 import kitchenpos.tablegroup.domain.TableGroup;
