+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -20,9 +20,9 @@ import kitchenpos.table.repository.OrderTableRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.common.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.domain.Price;
+import kitchenpos.common.Price;
 import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.order.dto.request.OrderLineItemCreateRequest;
 import kitchenpos.order.dto.request.OrderCreateRequest;
