+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.product.repository.ProductRepository;
-import kitchenpos.domain.Price;
+import kitchenpos.common.Price;
 import kitchenpos.product.doamin.Product;
 import kitchenpos.product.dto.request.ProductCreateRequest;
 import kitchenpos.product.dto.response.ProductResponse;
