+++ src/test/java/kitchenpos/ui/ProductRestControllerTest.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.product.ui.ProductRestController;
 import kitchenpos.product.application.ProductService;
-import kitchenpos.domain.Price;
+import kitchenpos.common.Price;
 import kitchenpos.product.doamin.Product;
 import kitchenpos.product.dto.request.ProductCreateRequest;
 import kitchenpos.product.dto.response.ProductResponse;
